General Update/New code
General design update, with minor fix issues. Added in rollers to reduce friction from the pins on the hose. This is represented by a new parameter in the F3d file. The rollers currently being used are nylon spacers for a #4 screw, as they are readily available. 
The Steps per uL calculator was also created. This takes the hose dimensions, and pump dimensions such as the hose bend radius and outputs what the steps/mm should be. 
Next steps are getting the Marlin fork to work to support five "axes"